+++ src/api/hooks/usePostOrder.ts
@@ -1,30 +1,10 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { fetchInstance } from '../instance';
-
-export interface OrderRequestBody {
-  productId: string;
-  productOptionId: number;
-  productQuantity: number;
-  messageCardTemplateId: number;
-  messageCardTextMessage: string;
-  senderId: number;
-  receiverId: number;
-  hasCashReceipt: boolean;
-  cashReceiptType: 'PERSONAL' | 'BUSINESS';
-  cashReceiptNumber: string;
-}
-
-const postOrderPath = () => '/v1/themes';
-const orderQueryKey = [postOrderPath()];
-
-export const postOrder = async (body: OrderRequestBody) => {
-  const response = await fetchInstance.post(postOrderPath(), body);
-  return response.data;
-};
+import { postOrder } from '../apis/order';
+import API from '../path';
 
 export const usePostOrder = () =>
   useMutation({
-    mutationKey: orderQueryKey,
+    mutationKey: [API.ORDER.POST_ORDER],
     mutationFn: postOrder,
   });
